Move roundTrip test to JTH
